requirement.txt
Django==2.2.13
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
psycopg2
gunicorn==20.0.4